// app/db/types/Pokemon.ts
export type Pokemon = {
  id: number;
  species_id: number;
  name_en: string;
  name_fr: string;
  name_ja: string;
  base_experience: number;
  height: number;
  order: number;
  weight: number;
  base_happiness: number;
  capture_rate: number;
  gender_rate: number;
  is_default_form: boolean;
  stat_hp: number;
  stat_attack: number;
  stat_defense: number;
  stat_special_attack: number;
  stat_special_defense: number;
  stat_speed: number;
  classification: string;
  color_id: number;
  shape_id: number;
  type_1_id: number;
  type_2_id: number;
  artwork_url: string;
  sprite_url: string;
  shiny_sprite_url: string;
};